refactor: definition 노드 리팩토링
