+++ agent/editor/nodes/definition.py
@@ -9,10 +9,14 @@ from agent.constants import (
     CATEGORY_TO_FILE,
     CATEGORY_TO_ID_FIELD,
     CATEGORY_TO_PLURAL,
-    PROPERTY_TO_FIELD,
 )
 from agent.core.llm_client import invoke_llm
-from agent.editor.nodes.game_index_resolve import apply_index_resolution
+from agent.editor.operation_ir import (
+    apply_index_resolution,
+    build_from_extractions,
+    normalize_modifications,
+    validate_operation_tuples,
+)
 from agent.editor.prompts.definition_prompt import (
     build_step1_prompt,
     build_step2_prompt,
@@ -417,88 +421,6 @@ def _format_to_progress_spec(modifications: list[dict], classifications: list[di
     return formatted_mods
 
 
-def _validate_operation_tuples(ops: list[dict], extractions: list[dict]) -> tuple[bool, str]:
-    """update/delete/read op의 degenerate 패턴 감지.
-
-    degenerate 판정 기준 (아래 중 하나 이상):
-      (a) action이 update/delete/read인데 subject에 id/name 모두 없음 (대상 미식별)
-      (b) action에 'update' 포함인데 updates 페이로드가 어디에도 비어있음
-
-    예: `독 상태이상의 지속 턴을 5로` → Step 4 매칭 실패로 subject.name=None →
-    (a) 적중. Executor에서 `update_all({})` silent no-op로 통과하는 패턴 차단.
-
-    Returns:
-        (True, "") — 정상
-        (False, message) — degenerate. message는 사용자 안내용
-    """
-    for op in ops:
-        action = (op.get("op") or op.get("action") or "").lower()
-        if not any(kw in action for kw in ("update", "delete", "read")):
-            continue
-
-        subject = op.get("subject") if isinstance(op.get("subject"), dict) else {}
-        sid = subject.get("id")
-        sname = subject.get("name")
-        # subject.name이 literal 'None' 문자열로 들어오는 케이스도 취급
-        if isinstance(sname, str) and sname.strip().lower() in ("", "none"):
-            sname = None
-
-        subject_missing = sid in (None, "", "NEW") and not sname
-
-        # updates 페이로드 탐색 (update 계열만)
-        updates_empty = False
-        if "update" in action:
-            # Step 5 경로: op["field"] + op["value"] 직접 구조
-            if op.get("field") and op.get("value") is not None:
-                updates_empty = False
-            else:
-                # Step 6 LLM 경로: target_info.updates / value.raw_updates
-                ti = op.get("target_info") or {}
-                updates = ti.get("updates") if isinstance(ti, dict) else None
-                if not updates:
-                    val = op.get("value") or {}
-                    if isinstance(val, dict):
-                        updates = val.get("raw_updates")
-                updates_empty = not isinstance(updates, dict) or len(updates) == 0
-
-        if not subject_missing and not updates_empty:
-            continue
-
-        # 안내 메시지 작성
-        hinted_name = sname or ""
-        if not hinted_name:
-            for ext in extractions:
-                ext_action = (ext.get("action") or "").upper()
-                if ext_action in ("UPDATE", "DELETE", "READ") and ext.get("subject"):
-                    hinted_name = ext["subject"]
-                    break
-
-        target_file = op.get("file") or "?"
-        if subject_missing and hinted_name:
-            msg = (
-                f"'{hinted_name}'을(를) {target_file}에서 찾지 못했습니다. "
-                "정확한 이름이나 ID를 알려주세요."
-            )
-        elif subject_missing:
-            msg = (
-                f"{target_file}에서 수정할 대상을 특정하지 못했습니다. "
-                "어떤 항목을 바꾸고 싶은지 구체적으로 알려주세요."
-            )
-        elif hinted_name:
-            msg = (
-                f"'{hinted_name}'에서 변경할 필드가 명확하지 않습니다. "
-                "어떤 값을 어떻게 바꿀지 알려주세요."
-            )
-        else:
-            msg = (
-                f"{target_file}에서 변경할 필드가 명확하지 않습니다. "
-                "어떤 값을 바꾸고 싶은지 알려주세요."
-            )
-        return False, msg
-
-    return True, ""
-
-
 async def definition(state: AgentState) -> dict:
     """사용자 입력을 operation IR 로 변환하는 10 단계 파이프라인.
 
@@ -792,7 +714,7 @@ async def definition(state: AgentState) -> dict:
         except Exception as _e:
             logger.warning("[Step5] 샘플맵 재랭킹 실패: %s → 기존 후보 유지", _e)
 
-    direct_ops = _extractions_to_operation_tuples(
+    direct_ops = build_from_extractions(
         extractions,
         final_classifications,
         final_extracted_ids,
@@ -808,7 +730,7 @@ async def definition(state: AgentState) -> dict:
         direct_ops = apply_index_resolution(direct_ops, game_id)
 
         # Step 10: degenerate op 검증 (update/update_all with empty updates → params_sufficient=False)
-        ok, reason = _validate_operation_tuples(direct_ops, extractions)
+        ok, reason = validate_operation_tuples(direct_ops, extractions)
         if not ok:
             logger.info("[Definition] Step 5 검증 실패(degenerate): %s", reason)
             return {
@@ -1025,7 +947,7 @@ async def definition(state: AgentState) -> dict:
     logger.info("[Definition] 노드 완료 - 생성된 modification 수: %d", len(strictly_formatted_mods))
 
     # operation IR 변환 (planner 가 소비)
-    operation_tuples = _to_operation_ir(strictly_formatted_mods, final_extracted_ids)
+    operation_tuples = normalize_modifications(strictly_formatted_mods, final_extracted_ids)
     logger.info("[Definition] operation IR 변환 완료: %d tuples", len(operation_tuples))
 
     # Step 9: GameIndex 기반 operation IR 후처리 (id/file/ref 확정, 장비 type_hint 보정)
@@ -1034,7 +956,7 @@ async def definition(state: AgentState) -> dict:
     # Step 10: degenerate op 검증 (update/update_all with empty updates → params_sufficient=False)
     degenerate_response: str | None = None
     if resolved_params_sufficient:
-        ok, reason = _validate_operation_tuples(operation_tuples, extractions)
+        ok, reason = validate_operation_tuples(operation_tuples, extractions)
         if not ok:
             logger.info("[Definition] 검증 실패(degenerate): %s", reason)
             resolved_params_sufficient = False
@@ -1062,559 +984,3 @@ async def definition(state: AgentState) -> dict:
         resolved_params_sufficient,
     )
     return result
-
-
-# ──────────────────────────────────────────────
-# Operation IR 변환 — modifications → operation_tuples
-# planner 가 소비하는 정규화된 IR 형식
-# ──────────────────────────────────────────────
-
-# _TARGET_TO_FILE, ARRAY_FIELDS, TRAIT_CODE_TO_HINT, CATEGORY_TO_FILE
-# → agent.constants 의 CATEGORY_TO_FILE, ARRAY_FIELDS, TRAIT_CODE_TO_HINT 로 통합
-# System.json 전용 확장
-_TARGET_TO_FILE: dict[str, str] = {
-    **CATEGORY_TO_FILE,
-    "element": "System.json",
-    "system": "System.json",
-}
-
-
-# ──────────────────────────────────────────────
-# Step 5: 코드 기반 operation IR 직접 생성
-# Step 1~4 결과(extractions + classifications)에서 LLM 없이 operation_tuples 생성
-# ──────────────────────────────────────────────
-
-# extraction.property → (field, value.kind) 매핑
-# property 가 이 테이블에 있으면 코드 기반 변환 가능
-# PROPERTY_TO_FIELD 는 agent.constants 에서 import
-
-
-def _extractions_to_operation_tuples(
-    extractions: list[dict],
-    classifications: list[dict],
-    extracted_ids: dict,
-    game_id: str,
-    ranked_map_candidates: list[dict] | None = None,
-) -> list[dict]:
-    """Step 1~4 결과에서 직접 operation_tuples 를 생성한다.
-
-    모든 extraction 을 변환할 수 없으면 빈 리스트를 반환 (fallback 신호).
-
-    ranked_map_candidates: 게임 생성 시 계산된 맵 후보군 (점수 순 정렬).
-        맵 추가 요청 시 이 목록에서 다음 후보를 선택해 original_file_name 을 주입한다.
-        None 또는 빈 리스트이면 맵 CREATE 는 Step 5(LLM) 로 fallback.
-
-    Returns:
-        list[dict] — 성공 시 operation_tuples, 실패 시 [].
-    """
-    # classification 을 이름 기반 lookup 으로 구성
-    cls_by_name: dict[str, dict] = {}
-    for cls in classifications:
-        name = cls.get("name", "").strip()
-        if name:
-            cls_by_name[name.lower()] = cls
-
-    logger.debug(
-        "[Step5] cls_by_name keys=%s, extractions=%d",
-        list(cls_by_name.keys()),
-        len(extractions),
-    )
-
-    ops: list[dict] = []
-
-    for ext in extractions:
-        action = (ext.get("action") or "").upper().strip()
-        subject = (ext.get("subject") or "").strip()
-        prop = (ext.get("property") or "").strip()
-        value_str = (ext.get("value") or "").strip()
-
-        if not subject:
-            return []  # 변환 불가
-
-        # subject 의 분류 정보 찾기 — 완전 일치 → 부분 포함
-        sub_cls = cls_by_name.get(subject.lower())
-        if not sub_cls:
-            # 부분 일치 fallback (예: subject="리드" vs cls name="리드 액터")
-            for cname, cval in cls_by_name.items():
-                if subject.lower() in cname or cname in subject.lower():
-                    sub_cls = cval
-                    break
-        if not sub_cls:
-            # System.json 수정: subject 가 카테고리 지칭어("게임", "시스템" 등)로 필터된 경우
-            if prop and value_str and action == "UPDATE":
-                sys_field = _detect_system_field(subject, prop)
-                if sys_field:
-                    logger.info(
-                        "[Step5] System.json 수정 감지: %s.%s = %s", subject, sys_field, value_str
-                    )
-                    ops.append(
-                        {
-                            "op": "update",
-                            "file": "System.json",
-                            "subject": None,
-                            "field": sys_field,
-                            "value": {
-                                "kind": "string",
-                                "ref": None,
-                                "new_value": value_str,
-                                "type_hint": None,
-                                "array_op": None,
-                                "match_hint": None,
-                            },
-                        }
-                    )
-                    continue
-            logger.info(
-                "[Step5] subject '%s' not in cls_by_name %s → fallback",
-                subject,
-                list(cls_by_name.keys()),
-            )
-            return []
-
-        sub_cat = (sub_cls.get("category") or "").lower()
-        sub_id = sub_cls.get("mapped_id")
-        if isinstance(sub_id, str):
-            sub_id = None if sub_id == "NEW" else int(sub_id) if sub_id.isdigit() else None
-
-        # extracted_ids 에서도 보완
-        if sub_id is None:
-            sub_id = extracted_ids.get(subject)
-
-        sub_file = CATEGORY_TO_FILE.get(sub_cat)
-        if not sub_file and sub_cat not in ("element", "system", "none"):
-            return []
-
-        # CREATE
-        if action == "CREATE":
-            if sub_cat == "map":
-                # 맵 추가: ranked_map_candidates 가 있으면 최상위 후보를 사용.
-                # 없으면 Step 5(LLM) 경로로 fallback.
-                if ranked_map_candidates:
-                    best = ranked_map_candidates[0]
-                    original_file_name = best["entry"]["file_name"]
-                    logger.info(
-                        "[Step5] 맵 추가 요청 → 순위 후보 사용: %s (score=%.3f)",
-                        original_file_name,
-                        best.get("score", 0.0),
-                    )
-                    ops.append(
-                        {
-                            "op": "create",
-                            "file": sub_file or "MapInfos.json",
-                            "subject": {"name": subject, "id": None, "scope": "single"},
-                            "field": None,
-                            "value": {
-                                "kind": "map_creation",
-                                "original_file_name": original_file_name,
-                            },
-                        }
-                    )
-                    continue
-                else:
-                    # 후보군 없음 → Step 5(LLM)에서 처리
-                    logger.info(
-                        "[Step5] 맵 추가 요청이지만 ranked_map_candidates 없음 → Step 5 fallback"
-                    )
-                    return []
-            ops.append(
-                {
-                    "op": "create",
-                    "file": sub_file or "System.json",
-                    "subject": {"name": subject, "id": None, "scope": "single"},
-                    "field": None,
-                    "value": None,
-                }
-            )
-            continue
-
-        # DELETE
-        if action == "DELETE":
-            ops.append(
-                {
-                    "op": "delete",
-                    "file": sub_file or "System.json",
-                    "subject": {"name": subject, "id": sub_id, "scope": "single"},
-                    "field": None,
-                    "value": None,
-                }
-            )
-            continue
-
-        # READ
-        if action == "READ":
-            ops.append(
-                {
-                    "op": "read",
-                    "file": sub_file or "System.json",
-                    "subject": {"name": subject, "id": sub_id, "scope": "single"},
-                    "field": prop if prop else None,
-                    "value": None,
-                }
-            )
-            continue
-
-        # UPDATE — property 로 field 결정
-        if action == "UPDATE":
-            if not prop and not value_str:
-                return []  # 뭘 수정하는지 모름
-
-            # value 에 대한 분류 정보 (참조 엔티티일 수 있음) — 완전 → 부분 일치
-            val_cls = None
-            if value_str:
-                val_cls = cls_by_name.get(value_str.lower())
-                if not val_cls:
-                    for cname, cval in cls_by_name.items():
-                        if value_str.lower() in cname or cname in value_str.lower():
-                            val_cls = cval
-                            break
-
-            # property 가 없으면 value 의 카테고리로 추론
-            if not prop and val_cls:
-                val_cat = (val_cls.get("category") or "").lower()
-                prop = _infer_property_from_value_category(val_cat)
-
-            if not prop:
-                return []  # field 결정 불가
-
-            # property → field 매핑
-            field_info = _lookup_property_field(prop)
-            if not field_info:
-                return []
-
-            field, kind = field_info
-
-            # value 구성
-            ir_value = _build_direct_ir_value(
-                field,
-                kind,
-                value_str,
-                val_cls,
-                extracted_ids,
-                prop=prop,
-            )
-
-            ops.append(
-                {
-                    "op": "update",
-                    "file": sub_file,
-                    "subject": {"name": subject, "id": sub_id, "scope": "single"},
-                    "field": field,
-                    "value": ir_value,
-                }
-            )
-            continue
-
-        # 알 수 없는 action
-        return []
-
-    return ops
-
-
-def _lookup_property_field(prop: str) -> tuple[str, str] | None:
-    """property 문자열에서 field 매핑을 찾는다. 완전 일치 → 부분 일치."""
-    t = prop.lower().replace(" ", "")
-    # 완전 일치
-    for key, val in PROPERTY_TO_FIELD.items():
-        if key.replace(" ", "") == t:
-            return val
-    # 부분 포함
-    for key, val in PROPERTY_TO_FIELD.items():
-        if key in prop or prop in key:
-            return val
-    return None
-
-
-def _detect_system_field(subject: str, prop: str) -> str | None:
-    """subject + property 에서 System.json 필드를 감지. 해당하면 필드명 반환."""
-    combined = (subject + " " + prop).lower()
-    if "제목" in combined or "타이틀" in combined:
-        return "gameTitle"
-    if "통화" in combined or "화폐" in combined:
-        return "currencyUnit"
-    return None
-
-
-def _infer_property_from_value_category(val_cat: str) -> str:
-    """value 의 카테고리에서 property 를 추론."""
-    return {
-        "skill": "스킬",
-        "weapon": "무기",
-        "armor": "장비",
-        "class": "직업",
-        "item": "드롭",
-        "state": "속성",
-        "element": "속성",
-    }.get(val_cat, "")
-
-
-def _build_direct_ir_value(
-    field: str,
-    kind: str,
-    value_str: str,
-    val_cls: dict | None,
-    extracted_ids: dict,
-    prop: str = "",
-) -> dict:
-    """직접 변환용 IR value 생성."""
-    val_id = None
-    if val_cls:
-        mid = val_cls.get("mapped_id")
-        if isinstance(mid, int):
-            val_id = mid
-        elif isinstance(mid, str) and mid.isdigit():
-            val_id = int(mid)
-    if val_id is None and value_str:
-        val_id = extracted_ids.get(value_str)
-
-    # 참조 필드 (learnings, equips, classId, actions, dropItems)
-    if kind in ("skill", "weapon", "armor", "class", "item"):
-        return {
-            "kind": kind,
-            "ref": value_str if value_str else None,
-            "new_value": None,
-            "type_hint": prop if prop else None,  # "방패", "장신구" 등 슬롯 힌트
-            "array_op": None,
-            "match_hint": None,
-        }
-
-    # trait 관련
-    if kind == "trait":
-        return {
-            "kind": "trait",
-            "ref": value_str if value_str else None,
-            "new_value": None,
-            "type_hint": None,
-            "array_op": "update",
-            "match_hint": "공격 속성" if "속성" in field or "공격" in (value_str or "") else None,
-        }
-
-    # 단순 값 (string, param)
-    if kind == "param":
-        try:
-            num = int(value_str) if value_str else None
-        except ValueError:
-            try:
-                num = float(value_str)
-            except ValueError:
-                num = None
-        return {
-            "kind": "param",
-            "ref": None,
-            "new_value": num,
-            "type_hint": None,
-            "array_op": None,
-            "match_hint": None,
-        }
-
-    # string
-    return {
-        "kind": "string",
-        "ref": None,
-        "new_value": value_str if value_str else None,
-        "type_hint": None,
-        "array_op": None,
-        "match_hint": None,
-    }
-
-
-def _to_operation_ir(
-    modifications: list[dict],
-    extracted_ids: dict,
-) -> list[dict]:
-    """기존 modifications 를 operation_tuples (operation IR) 로 변환.
-
-    modification 형식 (참고):
-        {
-            "type": "create" | "update" | "delete" | "query",
-            "target": "actor" | "enemy" | ...,
-            "params": {
-                # create: {"name": ..., "description": ..., ...}
-                # update: {"selector": {...}, "updates": {...}}
-                # query: {"searchTerm": ...}
-            }
-        }
-
-    operation IR 형식:
-        {
-            "op": "create" | "update" | "delete" | "read",
-            "file": "Actors.json",
-            "subject": {"name": str|None, "id": int|None, "scope": "single"|"all"},
-            "field": str|None,
-            "value": {kind, ref, new_value, type_hint, array_op, match_hint} | None,
-        }
-    """
-    op_type_map = {
-        "create": "create",
-        "update": "update",
-        "delete": "delete",
-        "query": "read",
-        "read": "read",
-    }
-
-    result: list[dict] = []
-    for mod in modifications:
-        mod_type = (mod.get("type") or "").lower()
-        target = (mod.get("target") or "").lower()
-        params = mod.get("params") or {}
-
-        op = op_type_map.get(mod_type)
-        file = _TARGET_TO_FILE.get(target)
-        if not op or not file:
-            continue
-
-        if op == "create":
-            tuples = _create_to_ir(file, target, params, extracted_ids)
-        elif op == "update":
-            tuples = _update_to_ir(file, target, params, extracted_ids)
-        elif op == "delete":
-            tuples = _delete_to_ir(file, target, params, extracted_ids)
-        elif op == "read":
-            tuples = _read_to_ir(file, target, params)
-        else:
-            tuples = []
-
-        result.extend(tuples)
-
-    return result
-
-
-def _create_to_ir(
-    file: str,
-    target: str,
-    params: dict,
-    extracted_ids: dict,
-) -> list[dict]:
-    """create modification → operation IR.
-
-    profiler 가 세부 필드를 채우므로, IR 은 name 만 담는다.
-    맵 생성 시 original_file_name 이 params 에 있으면 value 에 포함한다.
-    """
-    name = params.get("name") or params.get(f"{target}_name") or ""
-    value = None
-    if target == "map" and "original_file_name" in params:
-        value = {"kind": "map_creation", "original_file_name": params["original_file_name"]}
-    return [
-        {
-            "op": "create",
-            "file": file,
-            "subject": {"name": name, "id": None, "scope": "single"} if name else None,
-            "field": None,
-            "value": value,
-        }
-    ]
-
-
-def _update_to_ir(
-    file: str,
-    target: str,
-    params: dict,
-    extracted_ids: dict,
-) -> list[dict]:
-    """update modification → operation IR.
-
-    세부 필드 값(params, traits, effects 등)은 profiler 책임이므로 여기서는
-    스칼라 필드(name, initialLevel 등)만 raw_updates 로 전달한다.
-    배열/복합 필드는 field 이름만 남기고 값은 버린다.
-    """
-    selector = params.get("selector") or {}
-    updates = params.get("updates") or {}
-
-    # selector 분해
-    if isinstance(selector, dict):
-        scope = "all" if selector.get("mode") == "all" else "single"
-        sub_name = selector.get("name")
-        sub_id = selector.get("id")
-    else:
-        scope = "single"
-        sub_name = params.get("name")
-        sub_id = None
-
-    # extracted_ids 에서 id 보완
-    if sub_id is None and sub_name:
-        sub_id = extracted_ids.get(sub_name)
-
-    subject = None
-    if scope == "all":
-        subject = {"name": None, "id": None, "scope": "all"}
-    elif sub_name or sub_id:
-        subject = {"name": sub_name, "id": sub_id, "scope": "single"}
-
-    if not updates or not isinstance(updates, dict):
-        return []
-
-    # profiler 가 채울 복합 필드(traits, effects, params 등)는 버리고,
-    # 스칼라 + 참조 필드(equips, classId 등)는 남긴다.
-    _PROFILER_FIELDS = {
-        "traits",
-        "effects",
-        "damage",
-        "params",
-        "actions",
-        "dropItems",
-        "learnings",
-    }
-    filtered_updates = {k: v for k, v in updates.items() if k not in _PROFILER_FIELDS}
-
-    if not filtered_updates:
-        return []
-
-    return [
-        {
-            "op": "update",
-            "file": file,
-            "subject": subject,
-            "field": None,
-            "value": {
-                "kind": "updates",
-                "ref": None,
-                "new_value": None,
-                "type_hint": None,
-                "array_op": None,
-                "match_hint": None,
-                "raw_updates": filtered_updates,
-            },
-        }
-    ]
-
-
-def _delete_to_ir(
-    file: str,
-    target: str,
-    params: dict,
-    extracted_ids: dict,
-) -> list[dict]:
-    selector = params.get("selector") or {}
-    if isinstance(selector, dict):
-        sub_name = selector.get("name") or params.get("name")
-        sub_id = selector.get("id")
-    else:
-        sub_name = params.get("name")
-        sub_id = None
-    if sub_id is None and sub_name:
-        sub_id = extracted_ids.get(sub_name)
-
-    return [
-        {
-            "op": "delete",
-            "file": file,
-            "subject": {"name": sub_name, "id": sub_id, "scope": "single"},
-            "field": None,
-            "value": None,
-        }
-    ]
-
-
-def _read_to_ir(file: str, target: str, params: dict) -> list[dict]:
-    name = params.get("searchTerm") or params.get("name")
-    return [
-        {
-            "op": "read",
-            "file": file,
-            "subject": {"name": name, "id": None, "scope": "single"} if name else None,
-            "field": params.get("property"),
-            "value": None,
-        }
-    ]
-
-
-# _build_ir_value 삭제 — _update_to_ir 에서 스칼라만 통과시키므로 불필요.
